requirements.txt
streamlit
pdf2image
pytesseract
pandas
openai
fuzzywuzzy
python-Levenshtein
scikit-learn
Pillow
matplotlib
python-dotenv
logging
fitz